Finished condiment feature
Created CondimentGlob prefab and CondimentLeader prefab. These are intended to work together: Condimentleader should fall on to an ingredient so it can interact with the CondimentGlob prefab. Once the CondimentGlob prefab touches an ingredient, it sticks to it. Then it waits for a second ingredient, and once it touches a second ingredient it connects both ingredients.