CondimentBottle Play area detection
Created a way for the condimentBottle prefab to detect if it is in the play area. If the bottle is within a trigger that is on layer "CondimentBottle" then it will consider itself in the play area. Once it leaves that trigger zone, it will consider itself out of the play area.